pyproject.toml
# PACKAGE METADATA
[project]
name = "ragerrorclassification"
version = "0.0.1"
description = "Error Classification for RAG"
authors = []

requires-python = ">=3.10"

dependencies = [
    "numpy>=1.26.4",
    "tqdm>=4.66.1",
    "transformers>=4.40.1",
    "ftfy>=6.2.0",
    "openai>=1.25.1",
    "torch>=2.3.0",
    "matplotlib>=3.8.4",
    "faiss-cpu==1.7.4",
    "sentencepiece>=0.2.0",
    "protobuf>=5.26.1",
    "httpx>=0.27.0",
    "line-profiler>=4.1.3",
    "hydra-core>=1.3",
    "evaluate>=0.4.3",
    "dataclasses-json>=0.6.7",
    "spacy>=3.8.2",
    "litellm>=1.52.6",
    "rouge-score>=0.1.2",
]

# DEV DEPENDENCIES
[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.2.2",
    "pytest-cov>=5.0.0",
    "diff-cover>=9.0.0",
    "ipdb>=0.13",
]

# PDM SCRIPTS
[tool.pdm.scripts]
unit_tests.cmd = "pytest --cov --cov-report xml tests"
unit_tests.help = "Run unit tests with coverage"

cov_check.cmd = "diff-cover coverage.xml --config-file pyproject.toml"
cov_check.help = "Check unit test coverage between current branch and 'origin/main'"

# UNITTEST CONFIG
[tool.pytest.ini_options]
pythonpath = [
  "."
]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_also = [
  # don't complain if tests don't hit defensive assertion code
  "raise NotImplementedError",
  # don't complain about abstract methods
  "@(abc\\.)?abstractmethod",
  # don't complain about overload methods (their body is supposed to be empty)
  "@(typing\\.)?overload",
]

[tool.ruff]
line-length = 100

[tool.diff_cover]
compare_branch = "origin/main"
